requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf-yasg==1.21.9
inflection==0.5.1
packaging==24.2
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1